+++ crates/api/src/term/is_compatible_with.rs
@@ -1,8 +1,6 @@
 use std::borrow::Cow;
 
-use crate::{
-    parser::annotation_composition::AnnotationComposable, Composable, IsCompatibleWith, Term,
-};
+use crate::{annotation_composition::AnnotationComposable, Composable, IsCompatibleWith, Term};
 
 /// In order to enforce compatibility on "non-units" (ex. `{each}`, `{total}`, `{heartbeats}`),
 /// `Term`s need to compare their annotations along with their `Composition`s. In practice, and